# examples/cellar/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Cellar Service","description":"HTTP service for managing your wine cellar","version":""},"host":"localhost:8080","paths":{"/sommelier":{"post":{"tags":["sommelier"],"summary":"pick sommelier","operationId":"sommelier#pick","parameters":[{"name":"PickRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/PickRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/StoredBottleResponseBodyCollection"}},"400":{"description":"Bad Request response.","schema":{"$ref":"#/definitions/PickNoCriteriaResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/PickNoMatchResponseBody"}}},"schemes":["http"]}},"/storage":{"get":{"tags":["storage"],"summary":"list storage","description":"List all stored bottles","operationId":"storage#list","responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/StoredBottleResponseBodyTinyCollection"}}},"schemes":["http"]},"post":{"tags":["storage"],"summary":"add storage","description":"Add new bottle and return its ID.","operationId":"storage#add","parameters":[{"name":"AddRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/AddRequestBody"}}],"responses":{"201":{"description":"Created response.","schema":{"type":"string"}}},"schemes":["http"]}},"/storage/multi_add":{"post":{"tags":["storage"],"summary":"multi_add storage","description":"Add n number of bottles and return their IDs. This is a multipart request and each part has field name 'bottle' and contains the encoded bottle info to be added.","operationId":"storage#multi_add","parameters":[{"name":"array","in":"body","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/BottleRequestBody"}}}],"responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"type":"string","example":"Est beatae eveniet et tempora nobis enim."}}}},"schemes":["http"]}},"/storage/multi_update":{"put":{"tags":["storage"],"summary":"multi_update storage","description":"Update bottles with the given IDs. This is a multipart request and each part has field name 'bottle' and contains the encoded bottle info to be updated. The IDs in the query parameter is mapped to each part in the request.","operationId":"storage#multi_update","parameters":[{"name":"ids","in":"query","description":"IDs of the bottles to be updated","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"},{"name":"MultiUpdateRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/MultiUpdateRequestBody"}}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/storage/rate":{"post":{"tags":["storage"],"summary":"rate storage","description":"Rate bottles by IDs","operationId":"storage#rate","responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/storage/{id}":{"get":{"tags":["storage"],"summary":"show storage","description":"Show bottle by ID","operationId":"storage#show","parameters":[{"name":"view","in":"query","description":"View to render","required":false,"type":"string","enum":["default","tiny"]},{"name":"id","in":"path","description":"ID of bottle to show","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/ShowResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/ShowNotFoundResponseBody"}}},"schemes":["http"]},"delete":{"tags":["storage"],"summary":"remove storage","description":"Remove bottle from storage","operationId":"storage#remove","parameters":[{"name":"id","in":"path","description":"ID of bottle to remove","required":true,"type":"string"}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/swagger/swagger.json":{"get":{"summary":"Download ../../gen/http/openapi.json","description":"JSON document containing the API swagger definition","operationId":"swagger#/swagger/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"AddRequestBody":{"title":"AddRequestBody","type":"object","properties":{"composition":{"type":"array","items":{"$ref":"#/definitions/ComponentRequestBody"},"description":"Composition is the list of grape varietals and associated percentage.","example":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}]},"description":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100},"rating":{"type":"integer","description":"Rating of bottle from 1 (worst) to 5 (best)","example":1,"minimum":1,"maximum":5},"vintage":{"type":"integer","description":"Vintage of bottle","example":1958,"minimum":1900,"maximum":2020},"winery":{"$ref":"#/definitions/WineryRequestBody"}},"example":{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":2,"vintage":1943,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},"required":["name","winery","vintage"]},"BottleRequestBody":{"title":"BottleRequestBody","type":"object","properties":{"composition":{"type":"array","items":{"$ref":"#/definitions/ComponentRequestBody"},"description":"Composition is the list of grape varietals and associated percentage.","example":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}]},"description":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100},"rating":{"type":"integer","description":"Rating of bottle from 1 (worst) to 5 (best)","example":4,"minimum":1,"maximum":5},"vintage":{"type":"integer","description":"Vintage of bottle","example":2015,"minimum":1900,"maximum":2020},"winery":{"$ref":"#/definitions/WineryRequestBody"}},"description":"Bottle describes a bottle of wine to be stored.","example":{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":2,"vintage":1935,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},"required":["name","winery","vintage"]},"Component":{"title":"Component","type":"object","properties":{"percentage":{"type":"integer","description":"Percentage of varietal in wine","example":93,"minimum":1,"maximum":100},"varietal":{"type":"string","description":"Grape varietal","example":"Syrah","pattern":"[A-Za-z' ]+","maxLength":100}},"example":{"percentage":48,"varietal":"Syrah"},"required":["varietal"]},"ComponentRequestBody":{"title":"ComponentRequestBody","type":"object","properties":{"percentage":{"type":"integer","description":"Percentage of varietal in wine","example":17,"minimum":1,"maximum":100},"varietal":{"type":"string","description":"Grape varietal","example":"Syrah","pattern":"[A-Za-z' ]+","maxLength":100}},"example":{"percentage":58,"varietal":"Syrah"},"required":["varietal"]},"ComponentResponseBody":{"title":"ComponentResponseBody","type":"object","properties":{"percentage":{"type":"integer","description":"Percentage of varietal in wine","example":99,"minimum":1,"maximum":100},"varietal":{"type":"string","description":"Grape varietal","example":"Syrah","pattern":"[A-Za-z' ]+","maxLength":100}},"example":{"percentage":34,"varietal":"Syrah"},"required":["varietal"]},"MultiUpdateRequestBody":{"title":"MultiUpdateRequestBody","type":"object","properties":{"bottles":{"type":"array","items":{"$ref":"#/definitions/BottleRequestBody"},"description":"Array of bottle info that matches the ids attribute","example":[{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}}]}},"example":{"bottles":[{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":95,"varietal":"Syrah"},{"percentage":95,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","name":"Blue's Cuvee","rating":1,"vintage":1978,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}}]}},"PickNoCriteriaResponseBody":{"title":"PickNoCriteriaResponseBody","type":"string","description":"Missing criteria","example":"Optio corporis quam delectus."},"PickNoMatchResponseBody":{"title":"PickNoMatchResponseBody","type":"string","description":"No bottle matched given criteria","example":"Exercitationem alias est dolores."},"PickRequestBody":{"title":"PickRequestBody","type":"object","properties":{"name":{"type":"string","description":"Name of bottle to pick","example":"Blue's Cuvee"},"varietal":{"type":"array","items":{"type":"string","example":"Exercitationem ducimus expedita in quam eos distinctio."},"description":"Varietals in preference order","example":["pinot noir","merlot","cabernet franc"]},"winery":{"type":"string","description":"Winery of bottle to pick","example":"longoria"}},"example":{"name":"Blue's Cuvee","varietal":["pinot noir","merlot","cabernet franc"],"winery":"longoria"}},"ShowNotFoundResponseBody":{"title":"ShowNotFoundResponseBody","type":"object","properties":{"id":{"type":"string","description":"ID of missing bottle","example":"Quaerat sint voluptate."},"message":{"type":"string","description":"Message of error","example":"bottle 1 not found"}},"description":"Bottle not found","example":{"id":"Sit qui ut delectus.","message":"bottle 1 not found"},"required":["message","id"]},"ShowResponseBody":{"title":"Mediatype identifier: application/vnd.cellar.stored-bottle; view=default","type":"object","properties":{"composition":{"type":"array","items":{"$ref":"#/definitions/Component"},"description":"Composition is the list of grape varietals and associated percentage.","example":[{"percentage":86,"varietal":"Syrah"},{"percentage":86,"varietal":"Syrah"}]},"description":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"id":{"type":"string","description":"ID is the unique id of the bottle.","example":"123abc"},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100},"rating":{"type":"integer","description":"Rating of bottle from 1 (worst) to 5 (best)","example":4,"minimum":1,"maximum":5},"vintage":{"type":"integer","description":"Vintage of bottle","example":1908,"minimum":1900,"maximum":2020},"winery":{"$ref":"#/definitions/Winery"}},"description":"ShowResponseBody result type (default view)","example":{"composition":[{"percentage":86,"varietal":"Syrah"},{"percentage":86,"varietal":"Syrah"},{"percentage":86,"varietal":"Syrah"},{"percentage":86,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":2,"vintage":2007,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},"required":["id","name","winery","vintage"]},"StoredBottleResponseBody":{"title":"Mediatype identifier: application/vnd.cellar.stored-bottle; view=default","type":"object","properties":{"composition":{"type":"array","items":{"$ref":"#/definitions/ComponentResponseBody"},"description":"Composition is the list of grape varietals and associated percentage.","example":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}]},"description":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"id":{"type":"string","description":"ID is the unique id of the bottle.","example":"123abc"},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100},"rating":{"type":"integer","description":"Rating of bottle from 1 (worst) to 5 (best)","example":4,"minimum":1,"maximum":5},"vintage":{"type":"integer","description":"Vintage of bottle","example":1939,"minimum":1900,"maximum":2020},"winery":{"$ref":"#/definitions/WineryResponseBody"}},"description":"A StoredBottle describes a bottle retrieved by the storage service. (default view)","example":{"composition":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":3,"vintage":2013,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},"required":["id","name","winery","vintage"]},"StoredBottleResponseBodyCollection":{"title":"Mediatype identifier: ","type":"array","items":{"$ref":"#/definitions/StoredBottleResponseBody"},"description":"StoredBottleResponseBodyCollection is the result type for an array of StoredBottleResponseBody (default view)","example":[{"composition":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":3,"vintage":1913,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":3,"vintage":1913,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":3,"vintage":1913,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}},{"composition":[{"percentage":41,"varietal":"Syrah"},{"percentage":41,"varietal":"Syrah"}],"description":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":"123abc","name":"Blue's Cuvee","rating":3,"vintage":1913,"winery":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"}}]},"StoredBottleResponseBodyTiny":{"title":"Mediatype identifier: ","type":"object","properties":{"id":{"type":"string","description":"ID is the unique id of the bottle.","example":"123abc"},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100},"winery":{"$ref":"#/definitions/WineryResponseBodyTiny"}},"description":"A StoredBottle describes a bottle retrieved by the storage service. (tiny view) (default view) (default view) (default view)","example":{"id":"123abc","name":"Blue's Cuvee","winery":{"name":"Longoria"}},"required":["id","name","winery"]},"StoredBottleResponseBodyTinyCollection":{"title":"Mediatype identifier: ","type":"array","items":{"$ref":"#/definitions/StoredBottleResponseBodyTiny"},"description":"StoredBottleResponseBodyTinyCollection is the result type for an array of StoredBottleResponseBodyTiny (default view)","example":[{"id":"123abc","name":"Blue's Cuvee","winery":{"name":"Longoria"}},{"id":"123abc","name":"Blue's Cuvee","winery":{"name":"Longoria"}},{"id":"123abc","name":"Blue's Cuvee","winery":{"name":"Longoria"}},{"id":"123abc","name":"Blue's Cuvee","winery":{"name":"Longoria"}}]},"Winery":{"title":"Mediatype identifier: winery; view=default","type":"object","properties":{"country":{"type":"string","description":"Country of winery","example":"USA","pattern":"(?i)[a-z '\\.]+"},"name":{"type":"string","description":"Name of winery","example":"Longoria"},"region":{"type":"string","description":"Region of winery","example":"Central Coast, California","pattern":"(?i)[a-z '\\.]+"},"url":{"type":"string","description":"Winery website URL","example":"http://www.longoriawine.com/","pattern":"(?i)^(https?|ftp)://[^\\s/$.?#].[^\\s]*$"}},"example":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"},"required":["name","region","country"]},"WineryRequestBody":{"title":"Mediatype identifier: winery; view=default","type":"object","properties":{"country":{"type":"string","description":"Country of winery","example":"USA","pattern":"(?i)[a-z '\\.]+"},"name":{"type":"string","description":"Name of winery","example":"Longoria"},"region":{"type":"string","description":"Region of winery","example":"Central Coast, California","pattern":"(?i)[a-z '\\.]+"},"url":{"type":"string","description":"Winery website URL","example":"http://www.longoriawine.com/","pattern":"(?i)^(https?|ftp)://[^\\s/$.?#].[^\\s]*$"}},"description":"WineryRequestBody result type (default view)","example":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"},"required":["name","region","country"]},"WineryResponseBody":{"title":"Mediatype identifier: winery; view=default","type":"object","properties":{"country":{"type":"string","description":"Country of winery","example":"USA","pattern":"(?i)[a-z '\\.]+"},"name":{"type":"string","description":"Name of winery","example":"Longoria"},"region":{"type":"string","description":"Region of winery","example":"Central Coast, California","pattern":"(?i)[a-z '\\.]+"},"url":{"type":"string","description":"Winery website URL","example":"http://www.longoriawine.com/","pattern":"(?i)^(https?|ftp)://[^\\s/$.?#].[^\\s]*$"}},"example":{"country":"USA","name":"Longoria","region":"Central Coast, California","url":"http://www.longoriawine.com/"},"required":["name","region","country"]},"WineryResponseBodyTiny":{"title":"Mediatype identifier: ","type":"object","properties":{"name":{"type":"string","description":"Name of winery","example":"Longoria"}},"description":"WineryResponseBodyTiny result type (default view) (default view)","example":{"name":"Longoria"},"required":["name"]}}}